src/ml_cluster/ml_scripts/vit_requirements.txt
# do this first: 
# sudo pip3 install --upgrade pip

setuptools-rust
transformers
torch
Pillow